Split large desktop UI and bridge modules
