+++ apps/desktop-tauri/src-tauri/src/bin/bridge_runner/http/protocol.rs
@@ -1,0 +1,107 @@
+use std::io::{Read, Write};
+use std::net::TcpStream;
+use std::time::Duration;
+
+use anyhow::{anyhow, Result};
+
+#[derive(Debug)]
+pub(super) struct HttpRequestData {
+    pub(super) method: String,
+    pub(super) path: String,
+    pub(super) body: String,
+}
+
+pub(super) struct HttpResponse {
+    pub(super) status: &'static str,
+    pub(super) content_type: &'static str,
+    pub(super) body: String,
+}
+
+impl HttpResponse {
+    pub(super) fn json_ok(body: String) -> Self {
+        Self {
+            status: "200 OK",
+            content_type: "application/json",
+            body,
+        }
+    }
+
+    pub(super) fn json_error(status: &'static str, error: &str) -> Self {
+        Self {
+            status,
+            content_type: "application/json",
+            body: serde_json::json!({ "error": error }).to_string(),
+        }
+    }
+}
+
+pub(super) fn read_http_request(stream: &mut TcpStream) -> Result<HttpRequestData> {
+    stream.set_read_timeout(Some(Duration::from_secs(2)))?;
+    let mut buffer = Vec::new();
+    let mut temp = [0_u8; 1024];
+    let header_end = loop {
+        let read = stream.read(&mut temp)?;
+        if read == 0 {
+            anyhow::bail!("connection closed before headers");
+        }
+        buffer.extend_from_slice(&temp[..read]);
+        if let Some(index) = find_subslice(&buffer, b"\r\n\r\n") {
+            break index + 4;
+        }
+    };
+    let header_text = String::from_utf8_lossy(&buffer[..header_end]);
+    let mut lines = header_text.split("\r\n").filter(|line| !line.is_empty());
+    let request_line = lines
+        .next()
+        .ok_or_else(|| anyhow!("missing request line"))?;
+    let mut content_length = 0_usize;
+    for line in lines.clone() {
+        if let Some((name, value)) = line.split_once(':') {
+            if name.eq_ignore_ascii_case("content-length") {
+                content_length = value.trim().parse().unwrap_or_default();
+            }
+        }
+    }
+    let mut parts = request_line.split_whitespace();
+    let method = parts
+        .next()
+        .ok_or_else(|| anyhow!("missing request method"))?
+        .to_string();
+    let path = parts
+        .next()
+        .ok_or_else(|| anyhow!("missing request path"))?
+        .to_string();
+    while buffer.len() < header_end + content_length {
+        let read = stream.read(&mut temp)?;
+        if read == 0 {
+            break;
+        }
+        buffer.extend_from_slice(&temp[..read]);
+    }
+    let body =
+        String::from_utf8_lossy(&buffer[header_end..buffer.len().min(header_end + content_length)])
+            .to_string();
+
+    Ok(HttpRequestData { method, path, body })
+}
+
+pub(super) fn write_http_response(
+    stream: &mut TcpStream,
+    status: &str,
+    content_type: &str,
+    body: &str,
+) -> Result<()> {
+    let response = format!(
+        "HTTP/1.1 {status}\r\nContent-Type: {content_type}\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{body}",
+        body.len()
+    );
+    stream.write_all(response.as_bytes())?;
+    stream.flush()?;
+    Ok(())
+}
+
+fn find_subslice(haystack: &[u8], needle: &[u8]) -> Option<usize> {
+    haystack
+        .windows(needle.len())
+        .position(|window| window == needle)
+}
